Altered to use HTTP rather than MQTT...
Which might have been a mistake. Anyway, both versions are still here and if I could be bothered I'd combine them with some kind of runtime or compile-time flag to control the comms method.

Also tidied up the logging, and in the HTTP case added Arduino/ESP-IDT lower level logging capability - uncomment the -D in the platformio.ini